Encapsulate all the code managing the per-view uniforms
This used to reside in View and it is now its own class. The reality is
that "per view" UBO is a bit of a misnomer -- currently we indeed have
one per View, but that's not required and in fact, in the future we might
want to have a separate one for shadow maps for instance.

A better name would be "per camera". Currently it is dynamically updated 
when rendering the View and shadow maps. In the future this probably need
to be separated.

+++ filament/src/Exposure.cpp
@@ -132,7 +132,7 @@ float exposure(float ev100) noexcept {
     // with the maximum luminance Lmax
     //
     // Reference: https://en.wikipedia.org/wiki/Film_speed
-    return static_cast<float>(1.0 / (1.2 * std::pow(2.0, ev100)));
+    return 1.0f / (1.2f * std::pow(2.0f, ev100));
 }
 
 float luminance(const Camera& c) noexcept {
